+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.9] / 2022-03-24
+### Fixed
+- Fix `resources.dll` problem, remove from `.addin`
+### Added
+- Add `CreateRevitAddinOnProjectFiles` on `IRevitPackageBuilder`
+- Add Exemple Resource `pt-BR`
+
 ## [1.1.8] / 2022-03-23
 ### Changed
 - Change `PackageBuilder`, Before `Sign`